requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
psycopg2-binary
